requirements.txt
Flask==0.10.1
Flask-Login==0.2.11
Flask-Mail==0.9.1
Flask-Mobility==0.1.1
Flask-OAuth==0.12
Flask-Principal==0.4.0
Flask-SQLAlchemy==2.0
Flask-Session==0.1.1
Flask-Social==1.6.2
Flask-WTF==0.10.3
Jinja2==2.7.3
Magic-file-extensions==0.2
MarkupSafe==0.23
MySQL-python==1.2.5
Pillow==2.4.0
SQLAlchemy==0.9.8
SQLAlchemy-Utils==0.27.8
Tempita==0.5.3dev
Werkzeug==0.9.6
blinker==1.3
decorator==3.4.0
facebook-sdk==1.0.0-alpha
gnureadline==6.3.3
google-api-python-client==1.3.1
httplib2==0.9
ipython==2.3.1
itsdangerous==0.24
libxml2-python==2.9.1
oauth2==1.5.211
oauth2client==1.4.2
oauthlib==0.7.2
passlib==1.6.2
pbr==0.10.0
pyasn1==0.1.7
pyasn1-modules==0.0.5
python-dateutil==2.2
python-twitter==2.0
requests==2.4.3
requests-oauthlib==0.4.2
rsa==3.1.4
simplejson==3.6.5
six==1.8.0
sqlalchemy-migrate==0.9.2
sqlparse==0.1.13
uritemplate==0.6
wsgiref==0.1.2
